The resource/result.py found its way back into the repo after being deleted (probably in merge commit e24a3a8). Now it is deleted again and all changes have been ported to run.py
